compress the settings
Compress the programs and formatters together. This reduces quite a bit
of complexity in the formatter folder and also makes formatter settings
more directly available.

The main downside is that it's not possible to extend the programs
schema with custom settings anymore (see rustfmt and ormolu). The other
downsides are that the settings are not cleanly separated and programs
not cleanly exposed anymore (ie: no defaultText).