fix: only eval enable option if visible
When using `lib.mkRenamedOptionModule`, the trace warning is emitted
whenever the option is accessed.

Because we iterate and check the .enable option of all the programs, the
renamed option gets accessed all the time.

So we check if the option is "visible" before checking in, which
`lib.mkRenamedOptionModule` sets to false.

Fixes #211