nixpkgs-fmt: add `includes` and `excludes` options
It is as simple as the summary says.

I use `nixpkgs-fmt` in my Nix projects and the treefmt check for it
generally fails in system configuration projects because of generated
`hardware-configuration.nix` files not being formatted correctly, which
are meant to not be modified manually.

This makes it straightforward to add exclusions, rather than having to
go through `settings.formatter.nixpkgs-fmt.excludes` to do so.