Add support for rstfmt (#359)
* feat(rstfmt): adds support for rstfmt

Adds support for "rstfmt is a tool for automatically formatting reStructuredText files in a
consistent way.

* fix(rstfmt): add line_length option

To preserve the formatter defaults, a `line_length` option has been added. This enables users to
specify if they want a line_length other than 72, since that is controversially short in the modern
era.